Telegram: add test to README + rename file
