+++ src/bot/mining/closeMines.py
@@ -6,7 +6,7 @@ of a given user
 from src.common.logger import logger
 from src.common.txLogger import txLogger, logTx
 from src.helpers.sms import sendSms
-from src.helpers.instant_message import sendIM
+from src.helpers.instantMessage import sendIM
 from src.common.clients import crabadaWeb3Client
 from src.helpers.mines import (
     fetchOpenMines,
