+++ src/bot/mining/sendTeamsMining.py
@@ -5,7 +5,7 @@ Send a user's available teams mining
 from src.common.logger import logger
 from src.common.txLogger import txLogger, logTx
 from src.helpers.sms import sendSms
-from src.helpers.instant_message import sendIM
+from src.helpers.instantMessage import sendIM
 from src.common.clients import crabadaWeb3Client
 from src.helpers.teams import fetchAvailableTeamsForTask
 from src.models.User import User
