+++ src/tests/testSendIM.py
@@ -1,6 +1,6 @@
 """Code from https://www.twilio.com/docs/sms/quickstart/python#install-python-and-the-twilio-helper-library"""
 
-from src.helpers.instant_message import sendIM
+from src.helpers.instantMessage import sendIM
 from pprint import pprint
 
 # VARS
